+++ siad/core_test.go
@@ -1,5 +1,6 @@
-package sia
+package main
 
+/*
 import (
 	"math/big"
 	"testing"
@@ -84,3 +85,44 @@ func TestEverything(t *testing.T) {
 	// sendManyTransactions(t, c)
 	// testMinerDeadlocking(t, c)
 }
+
+func (c *Core) ScanMutexes() {
+	var state, host, hostdb, miner, renter, wallet int
+	go func() {
+		c.state.Height()
+		state++
+	}()
+	go func() {
+		c.host.HostInfo()
+		host++
+	}()
+	go func() {
+		c.hostDB.Size()
+		hostdb++
+	}()
+	go func() {
+		c.miner.Threads()
+		miner++
+	}()
+	go func() {
+		c.renter.RentInfo()
+		renter++
+	}()
+	go func() {
+		c.wallet.Balance(false)
+		wallet++
+	}()
+
+	go func() {
+		fmt.Println("mutex testing has started.")
+		time.Sleep(time.Second * 10)
+		fmt.Println("mutext testing results (0 means deadlock, 1 means success):")
+		fmt.Println("State: ", state)
+		fmt.Println("Host: ", host)
+		fmt.Println("HostDB: ", hostdb)
+		fmt.Println("Miner: ", miner)
+		fmt.Println("Renter: ", renter)
+		fmt.Println("Wallet: ", wallet)
+	}()
+}
+*/
